Move Unit impls to centralized place
