+++ signer/src/testing/dummy.rs
@@ -14,7 +14,12 @@ use bitcoin::TxOut;
 use bitvec::array::BitArray;
 use blockstack_lib::chainstate::{nakamoto, stacks};
 use clarity::util::secp256k1::Secp256k1PublicKey;
+use fake::Dummy;
 use fake::Fake;
+use fake::Faker;
+use p256k1::point::Point;
+use p256k1::scalar::Scalar;
+use polynomial::Polynomial;
 use rand::seq::IteratorRandom as _;
 use rand::Rng;
 use sbtc::deposits::DepositScriptInputs;
@@ -24,6 +29,27 @@ use secp256k1::SECP256K1;
 use stacks_common::address::AddressHashMode;
 use stacks_common::address::C32_ADDRESS_VERSION_TESTNET_MULTISIG;
 use stacks_common::types::chainstate::StacksAddress;
+use wsts::common::Nonce;
+use wsts::common::PolyCommitment;
+use wsts::common::PublicNonce;
+use wsts::common::SignatureShare;
+use wsts::common::TupleProof;
+use wsts::net::BadPrivateShare;
+use wsts::net::DkgBegin;
+use wsts::net::DkgEnd;
+use wsts::net::DkgEndBegin;
+use wsts::net::DkgFailure;
+use wsts::net::DkgPrivateBegin;
+use wsts::net::DkgPrivateShares;
+use wsts::net::DkgPublicShares;
+use wsts::net::DkgStatus;
+use wsts::net::NonceRequest;
+use wsts::net::NonceResponse;
+use wsts::net::SignatureShareRequest;
+use wsts::net::SignatureShareResponse;
+use wsts::net::SignatureType;
+use wsts::traits::PartyState;
+use wsts::traits::SignerState;
 
 use crate::bitcoin::utxo::Fees;
 use crate::bitcoin::validation::TxRequestIds;
@@ -758,5 +784,361 @@ impl fake::Dummy<fake::Faker> for BitcoinPreSignRequest {
 impl fake::Dummy<fake::Faker> for BitcoinPreSignAck {
     fn dummy_with_rng<R: rand::RngCore + ?Sized>(_config: &fake::Faker, _rng: &mut R) -> Self {
         BitcoinPreSignAck {}
+    }
+}
+/// A struct to help with creating dummy values for testing
+pub struct Unit;
+
+impl Dummy<Unit> for bitcoin::OutPoint {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        let bytes: [u8; 32] = Faker.fake_with_rng(rng);
+        let txid = bitcoin::Txid::from_byte_array(bytes);
+        let vout: u32 = Faker.fake_with_rng(rng);
+        bitcoin::OutPoint { txid, vout }
+    }
+}
+
+impl Dummy<Unit> for RecoverableSignature {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        let private_key = PrivateKey::new(rng);
+        let msg = secp256k1::Message::from_digest([0; 32]);
+        private_key.sign_ecdsa_recoverable(&msg)
+    }
+}
+
+impl Dummy<Unit> for secp256k1::ecdsa::Signature {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        let private_key = PrivateKey::new(rng);
+        let msg = secp256k1::Message::from_digest([0; 32]);
+        private_key.sign_ecdsa(&msg)
+    }
+}
+
+impl Dummy<Unit> for StacksAddress {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        let public_key: PublicKey = Faker.fake_with_rng(rng);
+        let pubkey = public_key.into();
+        let mainnet: bool = Faker.fake_with_rng(rng);
+        StacksAddress::p2pkh(mainnet, &pubkey)
+    }
+}
+
+impl Dummy<Unit> for Scalar {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        let number: [u8; 32] = Faker.fake_with_rng(rng);
+        p256k1::scalar::Scalar::from(number)
+    }
+}
+
+impl Dummy<Unit> for Point {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        Point::from(config.fake_with_rng::<Scalar, R>(rng))
+    }
+}
+
+impl Dummy<Unit> for Polynomial<Scalar> {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, _: &mut R) -> Self {
+        Polynomial::new(
+            fake::vec![[u8; 32]; 0..=15]
+                .into_iter()
+                .map(p256k1::scalar::Scalar::from)
+                .collect(),
+        )
+    }
+}
+
+impl Dummy<Unit> for (u32, Scalar) {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        (Faker.fake_with_rng(rng), config.fake_with_rng(rng))
+    }
+}
+
+impl Dummy<Unit> for DkgBegin {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        DkgBegin {
+            dkg_id: Faker.fake_with_rng(rng),
+        }
+    }
+}
+impl Dummy<Unit> for DkgPrivateBegin {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        DkgPrivateBegin {
+            dkg_id: Faker.fake_with_rng(rng),
+            signer_ids: Faker.fake_with_rng(rng),
+            key_ids: Faker.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for Vec<(u32, hashbrown::HashMap<u32, Vec<u8>>)> {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, _: &mut R) -> Self {
+        fake::vec![u32; 0..16]
+            .into_iter()
+            .map(|v| (v, fake::vec![(u32, Vec<u8>); 0..16].into_iter().collect()))
+            .collect()
+    }
+}
+
+impl Dummy<Unit> for DkgPrivateShares {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        DkgPrivateShares {
+            dkg_id: Faker.fake_with_rng(rng),
+            signer_id: Faker.fake_with_rng(rng),
+            shares: config.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for DkgEndBegin {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        DkgEndBegin {
+            dkg_id: Faker.fake_with_rng(rng),
+            signer_ids: Faker.fake_with_rng(rng),
+            key_ids: Faker.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for TupleProof {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        TupleProof {
+            R: config.fake_with_rng(rng),
+            rB: config.fake_with_rng(rng),
+            z: config.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for BadPrivateShare {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        BadPrivateShare {
+            shared_key: config.fake_with_rng(rng),
+            tuple_proof: config.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for hashbrown::HashMap<u32, BadPrivateShare> {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        fake::vec![u32; 0..20]
+            .into_iter()
+            .map(|v| (v, config.fake_with_rng(rng)))
+            .collect()
+    }
+}
+
+impl Dummy<Unit> for hashbrown::HashSet<u32> {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, _: &mut R) -> Self {
+        fake::vec![u32; 0..20].into_iter().collect()
+    }
+}
+
+impl Dummy<Unit> for DkgStatus {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        match rng.gen_range(0..6usize) {
+            0 => DkgStatus::Success,
+            1 => DkgStatus::Failure(DkgFailure::BadState),
+            2 => DkgStatus::Failure(DkgFailure::MissingPublicShares(config.fake_with_rng(rng))),
+            3 => DkgStatus::Failure(DkgFailure::BadPublicShares(config.fake_with_rng(rng))),
+            4 => DkgStatus::Failure(DkgFailure::MissingPrivateShares(config.fake_with_rng(rng))),
+            5 => DkgStatus::Failure(DkgFailure::BadPrivateShares(config.fake_with_rng(rng))),
+            _ => unreachable!(),
+        }
+    }
+}
+
+impl Dummy<Unit> for DkgEnd {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        DkgEnd {
+            dkg_id: Faker.fake_with_rng(rng),
+            signer_id: Faker.fake_with_rng(rng),
+            status: config.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for SignatureType {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &Unit, rng: &mut R) -> Self {
+        match rng.gen_range(0..3usize) {
+            0 => SignatureType::Frost,
+            1 => SignatureType::Schnorr,
+            2 => SignatureType::Taproot(if rng.gen_bool(0.5) {
+                None
+            } else {
+                Some(Faker.fake_with_rng(rng))
+            }),
+            _ => unreachable!(),
+        }
+    }
+}
+
+impl Dummy<Unit> for NonceRequest {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        NonceRequest {
+            dkg_id: Faker.fake_with_rng(rng),
+            sign_id: Faker.fake_with_rng(rng),
+            sign_iter_id: Faker.fake_with_rng(rng),
+            message: Faker.fake_with_rng(rng),
+            signature_type: config.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for PublicNonce {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        PublicNonce {
+            D: config.fake_with_rng(rng),
+            E: config.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for NonceResponse {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        NonceResponse {
+            dkg_id: Faker.fake_with_rng(rng),
+            sign_id: Faker.fake_with_rng(rng),
+            sign_iter_id: Faker.fake_with_rng(rng),
+            signer_id: Faker.fake_with_rng(rng),
+            key_ids: Faker.fake_with_rng(rng),
+            nonces: fake::vec![(); 0..20]
+                .into_iter()
+                .map(|_| config.fake_with_rng(rng))
+                .collect(),
+            message: Faker.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for SignatureShareRequest {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        SignatureShareRequest {
+            dkg_id: Faker.fake_with_rng(rng),
+            sign_id: Faker.fake_with_rng(rng),
+            sign_iter_id: Faker.fake_with_rng(rng),
+            nonce_responses: fake::vec![(); 0..20]
+                .into_iter()
+                .map(|_| config.fake_with_rng(rng))
+                .collect(),
+            signature_type: config.fake_with_rng(rng),
+            message: Faker.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for SignatureShare {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        SignatureShare {
+            id: Faker.fake_with_rng(rng),
+            z_i: config.fake_with_rng(rng),
+            key_ids: Faker.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for SignatureShareResponse {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        SignatureShareResponse {
+            dkg_id: Faker.fake_with_rng(rng),
+            sign_id: Faker.fake_with_rng(rng),
+            sign_iter_id: Faker.fake_with_rng(rng),
+            signer_id: Faker.fake_with_rng(rng),
+            signature_shares: fake::vec![(); 0..20]
+                .into_iter()
+                .map(|_| config.fake_with_rng(rng))
+                .collect(),
+        }
+    }
+}
+
+impl Dummy<Unit> for Nonce {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        Nonce {
+            d: config.fake_with_rng(rng),
+            e: config.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for (u32, PartyState) {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        (
+            Faker.fake_with_rng(rng),
+            PartyState {
+                polynomial: config.fake_with_rng(rng),
+                private_keys: fake::vec![(); 0..20]
+                    .into_iter()
+                    .map(|_| config.fake_with_rng(rng))
+                    .collect(),
+                nonce: config.fake_with_rng(rng),
+            },
+        )
+    }
+}
+
+impl Dummy<Unit> for SignerState {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        SignerState {
+            id: Faker.fake_with_rng(rng),
+            key_ids: Faker.fake_with_rng(rng),
+            num_keys: Faker.fake_with_rng(rng),
+            num_parties: Faker.fake_with_rng(rng),
+            threshold: Faker.fake_with_rng(rng),
+            group_key: config.fake_with_rng(rng),
+            parties: fake::vec![(); 0..20]
+                .into_iter()
+                .map(|_| config.fake_with_rng(rng))
+                .collect(),
+        }
+    }
+}
+
+impl Dummy<Unit> for wsts::schnorr::ID {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        wsts::schnorr::ID {
+            id: config.fake_with_rng(rng),
+            kG: config.fake_with_rng(rng),
+            kca: config.fake_with_rng(rng),
+        }
+    }
+}
+
+impl Dummy<Unit> for PolyCommitment {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        PolyCommitment {
+            id: config.fake_with_rng(rng),
+            poly: fake::vec![(); 0..20]
+                .into_iter()
+                .map(|_| config.fake_with_rng(rng))
+                .collect(),
+        }
+    }
+}
+
+impl Dummy<Unit> for (u32, PolyCommitment) {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        (Faker.fake_with_rng(rng), config.fake_with_rng(rng))
+    }
+}
+
+impl Dummy<Unit> for DkgPublicShares {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        DkgPublicShares {
+            dkg_id: Faker.fake_with_rng(rng),
+            signer_id: Faker.fake_with_rng(rng),
+            comms: fake::vec![(); 0..20]
+                .into_iter()
+                .map(|_| config.fake_with_rng(rng))
+                .collect(),
+        }
+    }
+}
+
+impl Dummy<Unit> for BTreeMap<u32, DkgPublicShares> {
+    fn dummy_with_rng<R: rand::Rng + ?Sized>(config: &Unit, rng: &mut R) -> Self {
+        fake::vec![(); 0..20]
+            .into_iter()
+            .map(|_| (Faker.fake_with_rng(rng), config.fake_with_rng(rng)))
+            .collect()
     }
 }
